ゴールドのテストで作られたキャッシュがウォレットのテストに影響を与えることがあった不具合を修正
